+++ beetsplug/inline.py
@@ -18,7 +18,6 @@ import logging
 import traceback
 
 from beets.plugins import BeetsPlugin
-from beets.library import Item, Album
 from beets import config
 
 log = logging.getLogger('beets')
@@ -47,14 +46,6 @@ def _compile_func(body):
     eval(code, env)
     return env[FUNC_NAME]
 
-def _record(obj):
-    """Get a dictionary of values for an Item or Album object.
-    """
-    if isinstance(obj, Item):
-        return dict(obj.record)
-    else:
-        return dict(obj._record)
-
 def compile_inline(python_code):
     """Given a Python expression or function body, compile it as a path
     field function. The returned function takes a single argument, an
@@ -79,10 +70,8 @@ def compile_inline(python_code):
 
     if is_expr:
         # For expressions, just evaluate and return the result.
-        def _expr_func(obj):
-            if not isinstance(obj, Item):
-                return u''
-            values = _record(obj)
+        def _expr_func(item):
+            values = dict(item.record)
             try:
                 return eval(code, values)
             except Exception as exc:
@@ -91,10 +80,8 @@ def compile_inline(python_code):
     else:
         # For function bodies, invoke the function with values as global
         # variables.
-        def _func_func(obj):
-            if not isinstance(obj, Item):
-                return u''
-            func.__globals__.update(_record(obj))
+        def _func_func(item):
+            func.__globals__.update(item.record)
             try:
                 return func()
             except Exception as exc: